% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-cache.R
\name{CacheReporter}
\alias{CacheReporter}
\title{Test reporter: interactive progress bar of errors.}
\description{
`CacheReporter` is a version of [testthat::ProgressReporter] that also 
displays information about the number of cached tests. When loaded,
{testcache} sets this to be the default reporter by setting
`options(testthat.default_reporter = CacheReporter)`
}
\concept{reporters}
\section{Super classes}{
\code{\link[testthat:Reporter]{testthat::Reporter}} -> \code{\link[testthat:ProgressReporter]{testthat::ProgressReporter}} -> \code{CacheReporter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CacheReporter-start_context}{\code{CacheReporter$start_context()}}
\item \href{#method-CacheReporter-show_header}{\code{CacheReporter$show_header()}}
\item \href{#method-CacheReporter-show_status}{\code{CacheReporter$show_status()}}
\item \href{#method-CacheReporter-add_result}{\code{CacheReporter$add_result()}}
\item \href{#method-CacheReporter-end_reporter}{\code{CacheReporter$end_reporter()}}
\item \href{#method-CacheReporter-clone}{\code{CacheReporter$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".start_context"><a href='../../testthat/html/Reporter.html#method-Reporter-.start_context'><code>testthat::Reporter$.start_context()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_line"><a href='../../testthat/html/Reporter.html#method-Reporter-cat_line'><code>testthat::Reporter$cat_line()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_tight"><a href='../../testthat/html/Reporter.html#method-Reporter-cat_tight'><code>testthat::Reporter$cat_tight()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_context_if_started"><a href='../../testthat/html/Reporter.html#method-Reporter-end_context_if_started'><code>testthat::Reporter$end_context_if_started()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_file"><a href='../../testthat/html/Reporter.html#method-Reporter-end_file'><code>testthat::Reporter$end_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_test"><a href='../../testthat/html/Reporter.html#method-Reporter-end_test'><code>testthat::Reporter$end_test()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="local_user_output"><a href='../../testthat/html/Reporter.html#method-Reporter-local_user_output'><code>testthat::Reporter$local_user_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="rule"><a href='../../testthat/html/Reporter.html#method-Reporter-rule'><code>testthat::Reporter$rule()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_test"><a href='../../testthat/html/Reporter.html#method-Reporter-start_test'><code>testthat::Reporter$start_test()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="update"><a href='../../testthat/html/Reporter.html#method-Reporter-update'><code>testthat::Reporter$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="cr"><a href='../../testthat/html/ProgressReporter.html#method-ProgressReporter-cr'><code>testthat::ProgressReporter$cr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="end_context"><a href='../../testthat/html/ProgressReporter.html#method-ProgressReporter-end_context'><code>testthat::ProgressReporter$end_context()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="initialize"><a href='../../testthat/html/ProgressReporter.html#method-ProgressReporter-initialize'><code>testthat::ProgressReporter$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="is_full"><a href='../../testthat/html/ProgressReporter.html#method-ProgressReporter-is_full'><code>testthat::ProgressReporter$is_full()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="report_issues"><a href='../../testthat/html/ProgressReporter.html#method-ProgressReporter-report_issues'><code>testthat::ProgressReporter$report_issues()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="should_update"><a href='../../testthat/html/ProgressReporter.html#method-ProgressReporter-should_update'><code>testthat::ProgressReporter$should_update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="start_file"><a href='../../testthat/html/ProgressReporter.html#method-ProgressReporter-start_file'><code>testthat::ProgressReporter$start_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="start_reporter"><a href='../../testthat/html/ProgressReporter.html#method-ProgressReporter-start_reporter'><code>testthat::ProgressReporter$start_reporter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="status_data"><a href='../../testthat/html/ProgressReporter.html#method-ProgressReporter-status_data'><code>testthat::ProgressReporter$status_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheReporter-start_context"></a>}}
\if{latex}{\out{\hypertarget{method-CacheReporter-start_context}{}}}
\subsection{Method \code{start_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheReporter$start_context(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheReporter-show_header"></a>}}
\if{latex}{\out{\hypertarget{method-CacheReporter-show_header}{}}}
\subsection{Method \code{show_header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheReporter$show_header()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheReporter-show_status"></a>}}
\if{latex}{\out{\hypertarget{method-CacheReporter-show_status}{}}}
\subsection{Method \code{show_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheReporter$show_status(complete = FALSE, time = 0, pad = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheReporter-add_result"></a>}}
\if{latex}{\out{\hypertarget{method-CacheReporter-add_result}{}}}
\subsection{Method \code{add_result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheReporter$add_result(context, test, result)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheReporter-end_reporter"></a>}}
\if{latex}{\out{\hypertarget{method-CacheReporter-end_reporter}{}}}
\subsection{Method \code{end_reporter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheReporter$end_reporter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheReporter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CacheReporter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
