% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_that.R
\name{test_that}
\alias{test_that}
\title{Run a test}
\usage{
test_that(desc, code)
}
\arguments{
\item{desc, code}{Passed to [testthat::test_that()]}
}
\value{
When run interactively, returns `invisible(TRUE)` if all tests pass, 
  otherwise throws an error.
}
\description{
This is a drop-in replacement for [testthat::test_that()], which first
checks a cache file to see if there have been any changes to the code
for the test, or any functions which it directly or indirectly calls. If
there haven't been any changes, the test is skipped.
}
